flask-backend/requirements.txt
Flask==1.1.2
Flask_Cors==4.0.1
Pillow==10.4.0
py_mini_racer==0.6.0
pymongo==4.8.0
